requirements.txt
tensorflow~=2.3.0
tensorflow-gpu~=2.3.0
tensorflow-estimator
pandas~=1.3.4
numpy~=1.21.4
scipy
matplotlib
opencv-python~=4.5.4.60
opencv-contrib-python
imutils
sklearn~=0.0
keras~=2.4.3
scikit-image~=0.19.0
scikit-plot
pillow~=8.4.0
seaborn~=0.11.2
scikit-learn~=1.0.1